Add diff output and confirmation

+++ load_crontab_workstation.sh
@@ -21,13 +21,44 @@ then
   exit 3
 fi
 
-# for "ts" command
-#
-brew install moreutils
 
 if [ ! -e "$FILEHOME/crontab-workstation.txt" ]
 then
   echo "Can't find crontab-workstation.txt in $FILEHOME"
   exit 2
 fi
-cat "$FILEHOME/crontab-workstation.txt" | $command
+
+if ! which -s ts
+then
+  # for "ts" command
+  #
+  brew install moreutils
+fi
+
+
+tempfile1=$(mktemp)
+tempfile2=$(mktemp)
+
+#cat "$FILEHOME/crontab-workstation.txt" | $command
+
+
+$command -l > $tempfile1
+cat "$FILEHOME/crontab-workstation.txt" > $tempfile2
+if [[ ! "$(diff $tempfile1 $tempfile2)" ]]
+then
+  echo "No updates. Aborting."
+  exit 1
+fi
+diff $tempfile1 $tempfile2
+rm $tempfile1 $tempfile2
+echo -n "Diff ok? (y/n): "
+read ans
+if [[ "$ans" == "y" ]]
+then
+  echo "Loading crontab."
+  cat "$FILEHOME/crontab-workstation.txt" | $command
+else
+  echo Aborting.
+fi
+
+
